Done: Proxy api<->remix, jwt validation security check
