+++ packages/reverse-proxy/custom-config/reverse-proxy-njs/loader.js
@@ -1,0 +1,1 @@
+global.JWTverify = require('jsonwebtoken').verify;
