+++ packages/reverse-proxy/custom-config/reverse-proxy.conf
@@ -5,20 +5,20 @@ upstream api-service {
     server app:5000;
 }
 
-# js_path "/etc/nginx/conf.d/reverse-proxy-njs/";
-# js_import auth.js;
-js_include /etc/nginx/conf.d/reverse-proxy-njs/dist/main.js;
-js_set $proxy_url validateCookieJwt;
+js_include /etc/nginx/conf.d/reverse-proxy-njs/dist/bundle.js;
+js_set $validateJwt validateCookieJwt;
 
 server {
     listen 5000;
     location /remix/ {
 
-    # return 200 "${SECRET_OR_KEY}" ;
         if ($cookie_SESSION_TOKEN = "") {
-            # redirect to login page or application -> proxy_pass http://api-service;
+            return 301 /;
         }
 
+        if ($validateJwt = "false") {
+            return 301 /;
+        }
 
         proxy_set_header X-Real-IP $remote_addr;
         proxy_set_header Host $host;
@@ -29,11 +29,12 @@ server {
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection "upgrade";
         rewrite /remix/(.*) /$1 break;
-        proxy_pass $proxy_url;
+        proxy_pass $scheme://remix-service;
+
     }
 
 
     location / {
-        proxy_pass http://api-service;
+        proxy_pass $scheme://api-service;
     }
 }
